# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "operationId": "getAllProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of products returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all products",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/available/in-stock": {
      "get": {
        "operationId": "getAvailableProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of available products returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get available products (IN_STOCK)",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "getOneProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific product",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a product",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a product",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/sell": {
      "patch": {
        "operationId": "sellProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product sold successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sell a product",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/kpi/my-space/sales/monthly": {
      "post": {
        "operationId": "ResellerSaleKpi_getMonthlySales",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParamsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Monthly sales returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonthlySalesDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get monthly sales for a reseller",
        "tags": [
          "Reseller KPIs - Sales"
        ]
      }
    },
    "/kpi/my-space/sales/average-ticket": {
      "post": {
        "operationId": "ResellerSaleKpi_getAverageTicket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParamsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Average ticket value",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get average ticket for a reseller",
        "tags": [
          "Reseller KPIs - Sales"
        ]
      }
    },
...
  },
  "info": {
    "title": "Luxis API",
    "description": "System API for Luxis",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "serialNumber": {
            "type": "string",
            "example": "0424A-BR-BAB-001"
          },
          "modelId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "batchId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "unitCost": {
            "type": "string",
            "example": "100.00"
          },
          "salePrice": {
            "type": "string",
            "example": "150.00"
          },
          "status": {
            "type": "string",
            "enum": [
              "IN_STOCK",
              "ASSIGNED",
              "SOLD"
            ],
            "example": "IN_STOCK"
          }
        },
        "required": [
          "id",
          "serialNumber",
          "modelId",
          "batchId",
          "unitCost",
          "salePrice",
          "status"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "unitCost": {
            "type": "string",
            "description": "The unit cost of the product",
            "example": "100.00"
          },
          "salePrice": {
            "type": "string",
            "description": "The sale price of the product",
            "example": "100.00"
          },
          "status": {
            "type": "string",
            "description": "The status of the product",
            "example": "IN_STOCK"
          }
        }
      },
...
      "ParamsDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T00:00:00Z"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T00:00:00Z"
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "page": {
            "type": "number",
            "example": 1
          }
        }
      },
...
      "MonthlySalesDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "number",
            "example": 1,
            "description": "Month number (1-12)"
          },
          "countSales": {
            "type": "number",
            "example": 1,
            "description": "Number of sales"
          },
          "totalAmount": {
            "type": "string",
            "example": "1000",
            "description": "Total sales amount"
          },
          "year": {
            "type": "number",
            "example": 2024,
            "description": "Year of sales"
          }
        },
        "required": [
          "month",
          "countSales",
          "totalAmount"
        ]
      },
...
    }
  }
}